Revert "Switch to fish for user shell"
I love parts of fish, but it breaks a lot of scripting that I'm
accustomed to using on the command line. If someone makes something like
fish (with the shiny completion menus and whatnot) that is bash- or
POSIX-compliant, I would just *love* that.

Shorter: fish is cool, but scritpting in it is non-portable which means
that I cant use it.

This reverts commit 5bfe1f3b19dde8e597bf2d73da81721bfa10a236.